renamed formatVariable-mw to formatVariable
